# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Test API using GO, Gin and Gorm",
        "title": "Test API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/alternativas/{questao_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alternativas"
                ],
                "summary": "List alternatives for a question",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Question ID",
                        "name": "questao_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Alternativa"
                            }
                        }
                    }
                }
            }
        },
        "/alunos": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Busca alunos pelo Nome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do aluno a ser buscado",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Aluno"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Cria um novo aluno",
                "parameters": [
                    {
                        "description": "Objeto Aluno a ser criado",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                }
            }
        },
        "/alunos/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Busca um aluno pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do aluno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Deleta um aluno pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do aluno a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Edita um aluno pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do aluno a ser editado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto Aluno com as alterações",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                }
            }
        },
        "/professores": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Lista todos os professores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Professor"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Cria um novo professor",
                "parameters": [
                    {
                        "description": "Objeto Professor a ser criado",
                        "name": "professor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Professor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Professor"
                        }
                    }
                }
            }
        },
        "/professores/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Busca um professor pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do professor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Professor"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Deleta um professor pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do professor a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Edita um professor pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do professor a ser editado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto Professor com as alterações",
                        "name": "professor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Professor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Professor"
                        }
                    }
                }
            }
        },
        "/professores/{nome}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professores"
                ],
                "summary": "Busca professores pelo Nome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do professor a ser buscado",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Professor"
                            }
                        }
                    }
                }
            }
        },
        "/questoes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questoes"
                ],
                "summary": "Exibe todas as Questoes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Questao"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questoes"
                ],
                "summary": "Cria uma nova Questao",
                "parameters": [
                    {
                        "description": "Objeto Questao a ser criado",
                        "name": "questao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Questao"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Questao"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Alternativa": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "resposta": {
                    "type": "boolean"
                },
                "texto": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Aluno": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "data_nascimento": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "turma_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Professor": {
            "type": "object",
            "properties": {
                "contato": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "especializacao": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Questao": {
            "type": "object",
            "properties": {
                "alternativas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Alternativa"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "enunciado": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "materia_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}